Move properties to own file, clean up implementation, implement properties for plugdata
